==> Proyecto_No_1_steam/Books/1_limpieza.ipynb <==
import pandas as pd
import ast
import json
import gzip

### Funciones

#Funcion que lee los archivos JSON

def leer(ruta):
    """Esta funcion lee los archivos JSON y los transforma a DataFrame"""
    lista=[]
    with open(ruta, 'r', encoding='UTF-8') as f:
        for line in f.readlines():
            try:
                data=ast.literal_eval(line)
                lista.append(data)
            except ValueError as e:
                lista.append(line)
                continue

    return pd.DataFrame(lista)

# Análisis de Calidad de Datos en un DataFrame"
def datos(df):
    verif = {"nombre_campo": [], "tipo_dato": [], "%_No_Nulos": [], "%_Nulos": [], "Nulos": []}

    for columna in df.columns:
        porcentaje_no_nulos = (df[columna].count() / len(df)) * 100
        verif["nombre_campo"].append(columna)
        verif["tipo_dato"].append(df[columna].dtypes)
        verif["%_No_Nulos"].append(round(porcentaje_no_nulos, 2))
        verif["%_Nulos"].append(round(100 - porcentaje_no_nulos, 2))
        verif["Nulos"].append(df[columna].isnull().sum())

    df_info = pd.DataFrame(verif)
        
    return df_info

### Review

df_reviews=leer("../../../dato/australian_user_reviews.json")
df_reviews.head()

# Obtener los datos de la columna Review

lista_review=[]
for ir, fr in df_reviews["reviews"].items():
    df_n_r=pd.json_normalize(fr)
    df_n_r["user_id"]=df_reviews.iloc[ir]["user_id"]
    lista_review.append(df_n_r)

df_reviews2=pd.concat(lista_review)
df_reviews2.head()

#Limpieza de datos

df_reviews2.drop(['funny','last_edited','helpful'], axis=1, inplace=True)
df_reviews2 = df_reviews2.dropna(how='all', axis=1)
df_reviews2 = df_reviews2.drop_duplicates()
df_reviews2.head()

datos(df_reviews2)

### Items

#Dado el tamaño del archivo JSON no se subira a git, y se convertira a formato parquet
df_items=leer("../../../dato/australian_users_items.json")
df_items.head()

#Extraer los datos guardados en los diccionarios de la columna de items
lista_items=[]
for i, f in df_items["items"].items():
    df_n_i=pd.json_normalize(f)
    df_n_i["user_id"]=df_items.iloc[i]["user_id"]
    df_n_i["items_count"]=df_items.iloc[i]["items_count"]
    #df_n_i["items_count"]=df_items.iloc[i]["items_count"]
    #df_n_i["steam_id"]=df_items.iloc[i]["steam_id"]
    lista_items.append(df_n_i)

df_items2=pd.concat(lista_items)
df_items2.head(5)

df_items2.drop(['item_name', 'playtime_2weeks'], axis=1, inplace=True)
df_items2 = df_items2.dropna(how='all')
df_items2 = df_items2.drop_duplicates()
df_items2.head()

datos(df_items2)

### Games

df_games=pd.read_json("../../../dato/output_steam_games.json", lines=True)
df_games.tail(3)

df_games = df_games.dropna(how="all").reset_index(drop=True) 
df_games.head()

df_games.drop(['title', 'url', 'early_access', 'reviews_url','specs'], axis=1, inplace=True)
df_games.head()

df_games['price'] = df_games['price'].astype(str)

### Guardar archivos

df_items2.to_csv("../data/items.csv", index=False)
df_reviews2.to_parquet("../data/reviews.parquet", index=False)
df_games.to_csv("../data/games.csv", index=False)
