==> Proyecto_No_1_steam/Books/2_Ingenieria.ipynb <==
import pandas as pd
from textblob import TextBlob

review=pd.read_parquet("../data/reviews.parquet")
review.head(3)

# Función de Análisis de Sentimiento: Determinando Polaridad en Texto

def analisis_sentimiento(review):
    analisis = TextBlob(str(review))
    # Asigna un valor según el análisis de sentimiento
    if analisis.sentiment.polarity > 0.1: 
        return 2  # Positivo
    elif analisis.sentiment.polarity < -0.1:
        return 0  # Malo
    else:
        return 1  # Neutral

# Aplicar la función a la columna 'reviews' y crear la nueva columna 'sentiment_analysis'
review['sentiment_analysis'] = review['review'].apply(analisis_sentimiento)

review.head()

review['sentiment_analysis'].value_counts()

review.drop(['review'], axis=1, inplace=True)

review.head()

review.to_csv("../data/review2.csv", index=False)
