==> Proyecto_No_1_steam/Books/5_modelo.ipynb <==
import pandas as pd
from sklearn.metrics.pairwise import cosine_similarity

# ML Recomendacion_juego

#### Cargar el Conjunto de Datos

df1 = pd.read_csv("../data/fusionada.csv")
df1.head()

df1.to_parquet('data_fusionada.parquet', index=False)

#### Seleccionar Columnas Relevantes

columnas = ['app_name', 'item_id', 'clasificacion']
df1 = df1[columnas]
df1.head()

# Filtrar Filas no Deseadas
df1 = df1[df1['clasificacion'] != 'Pendiente de clasificación']

#### Crear Variables Dummy

df_dummies = pd.get_dummies(df1, columns=['clasificacion'], prefix="")
df_dummies.head()

# Agrupar y Sumar
df_agrupado = df_dummies.groupby(['item_id', 'app_name'], as_index=False).sum()
df_agrupado.head()

#### Calcular la Similitud del Coseno

similitudes = cosine_similarity(df_agrupado.iloc[:, 2:])

#### Encontrar y Ordenar Juegos Similares

id_producto = 282010
idx = df_agrupado.index[df_agrupado['item_id'] == id_producto].tolist()[0]
similarity_scores = list(enumerate(similitudes[idx]))
similarity_scores = sorted(similarity_scores, key=lambda x: x[1], reverse=True)

#### Generar Recomendaciones

indices_similares = [i[0] for i in similarity_scores[1:6]]
juegos_recomendados = df_agrupado.iloc[indices_similares]['app_name']
juegos_recomendados

# ML recomendacion_usuario

### Carga y preparar los datos

# Especificar las columnas relevantes para el análisis
columnas = ['user_id', 'item_id', 'recommend', 'app_name']

# Cargar el DataFrame solo con las columnas especificadas
df = pd.read_csv('../data/fusionada.csv', usecols=columnas)

# Convertir 'recommend' a valores numéricos (1 para True, 0 para False)
df['interaction'] = df['recommend'].astype(int)


df.head()

#### Crear la Matriz de Utilidad

matriz_utilidad = df.pivot_table(index='user_id', columns='item_id', values='interaction', fill_value=0)

#### Calcular la Similitud entre Usuarios

similitudes = cosine_similarity(matriz_utilidad)
similitudes_df = pd.DataFrame(similitudes, index=matriz_utilidad.index, columns=matriz_utilidad.index)



#### Identificar Usuarios Similares

# Verificar si el usuario existe en el DataFrame
user_id = 'InstigatorAU'
if user_id not in similitudes_df.index:
    print("El ID de usuario proporcionado no existe en la base de datos.")
else:
    # Ordenar a los usuarios por similitud y seleccionar los top 10 más similares (excluyendo al propio usuario)
    usuarios_similares = similitudes_df[user_id].sort_values(ascending=False)[1:11]

#### Generar Recomendaciones

top_n = 5
juegos_recomendados = set()

for usuario_similar in usuarios_similares.index:
    # Juegos que el usuario similar ha recomendado
    juegos_usuario_similar = set(matriz_utilidad.columns[(matriz_utilidad.loc[usuario_similar] > 0)])
    # Juegos que el usuario objetivo ya conoce
    juegos_usuario = set(matriz_utilidad.columns[(matriz_utilidad.loc[user_id] > 0)])
    # Actualizar las recomendaciones con juegos que el usuario similar ha recomendado pero el usuario objetivo no conoce
    juegos_recomendados.update(juegos_usuario_similar.difference(juegos_usuario))

# Limitar a los top N recomendados
juegos_recomendados = list(juegos_recomendados)[:top_n]

#### Obtener Nombres de los Juegos Recomendados

# Obtener los nombres de los juegos recomendados
nombres_juegos = df[df['item_id'].isin(juegos_recomendados)]['app_name'].drop_duplicates().tolist()

nombres_juegos
